+++ apps/backoffice/src/lib/be/services/business.ts
@@ -1,5 +1,4 @@
 /* eslint-disable @typescript-eslint/no-explicit-any */
-import { getConfiguration } from "@/config";
 import { HTTP_STATUS_NO_CONTENT } from "@/config/constants";
 import { MigrationData } from "@/generated/api/MigrationData";
 import { MigrationDelegateList } from "@/generated/api/MigrationDelegateList";
@@ -7,6 +6,7 @@ import { MigrationItemList } from "@/generated/api/MigrationItemList";
 import { ServiceList } from "@/generated/api/ServiceList";
 import { ServiceTopicList } from "@/generated/api/ServiceTopicList";
 import { sanitizedNextResponseJson } from "@/lib/be/sanitize";
+import { SubscriptionCollection } from "@azure/arm-apimanagement";
 import { readableReport } from "@pagopa/ts-commons/lib/reporters";
 import { NonEmptyString } from "@pagopa/ts-commons/lib/strings";
 import * as E from "fp-ts/lib/Either";
@@ -15,20 +15,21 @@ import * as TE from "fp-ts/lib/TaskEither";
 import { pipe } from "fp-ts/lib/function";
 import { NextRequest } from "next/server";
 
-import { BackOfficeUser, Institution } from "../../../../types/next-auth";
+import { BackOfficeUser } from "../../../../types/next-auth";
 import {
   ManagedInternalError,
   handleBadRequestErrorResponse,
   handleInternalErrorResponse,
 } from "../errors";
 import { retrieveInstitutionGroups } from "../institutions/business";
-import { getServiceList } from "./apim";
+import { getSubscriptions } from "./apim";
 import {
   IoServicesCmsClient,
   callIoServicesCms,
   getServiceTopics,
 } from "./cms";
 import {
+  retrieveAuthorizedServiceIds,
   retrieveLifecycleServices,
   retrievePublicationServices,
 } from "./cosmos";
@@ -56,6 +57,35 @@ interface PathParameters {
   serviceId?: string;
 }
 
+const retrieveSubscriptions = (
+  backofficeUser: BackOfficeUser,
+  limit: number,
+  offset: number,
+  serviceId?: string,
+): TE.TaskEither<Error, SubscriptionCollection> =>
+  backofficeUser.permissions.selcGroups &&
+  backofficeUser.permissions.selcGroups.length > 0
+    ? pipe(
+        backofficeUser.permissions.selcGroups,
+        retrieveAuthorizedServiceIds,
+        TE.chain((authzServiceIds) =>
+          getSubscriptions(
+            backofficeUser.parameters.userId,
+            limit,
+            offset,
+            authzServiceIds.filter(
+              (authzServiceId) => !serviceId || authzServiceId === serviceId,
+            ),
+          ),
+        ),
+      )
+    : getSubscriptions(
+        backofficeUser.parameters.userId,
+        limit,
+        offset,
+        serviceId,
+      );
+
 /**
  * @description This method Will retrieve the specified list of services partition for the given user
  *
@@ -66,14 +96,13 @@ interface PathParameters {
  */
 export const retrieveServiceList = async (
   nextRequest: NextRequest,
-  userId: string,
-  institution: Institution,
+  backofficeUser: BackOfficeUser,
   limit: number,
   offset: number,
   serviceId?: string,
 ): Promise<ServiceList> =>
   pipe(
-    getServiceList(userId, limit, offset, serviceId),
+    retrieveSubscriptions(backofficeUser, limit, offset, serviceId),
     TE.bindTo("apimServices"),
     TE.bind("serviceTopicsMap", (_) =>
       pipe(
@@ -86,7 +115,7 @@ export const retrieveServiceList = async (
         TE.tryCatch(
           () =>
             retrieveInstitutionGroups(
-              institution.id,
+              backofficeUser.institution.id,
               1000, // FIXME: workaround to get all groups in a single call
               0,
             ),
@@ -156,7 +185,7 @@ export const retrieveServiceList = async (
           ...missingServices.map((missingService) =>
             buildMissingService(
               missingService.id,
-              institution,
+              backofficeUser.institution,
               missingService.createdDate,
             ),
           ),
@@ -204,11 +233,10 @@ export async function forwardIoServicesCmsRequest<
       "x-subscription-id": backofficeUser.parameters.subscriptionId,
       "x-user-email": backofficeUser.parameters.userEmail,
       "x-user-groups": backofficeUser.permissions.apimGroups.join(","),
-      "x-user-groups-selc": getConfiguration().GROUP_AUTHZ_ENABLED
-        ? backofficeUser.permissions.selcGroups?.join(",")
-        : undefined,
+      "x-user-groups-selc":
+        backofficeUser.permissions.selcGroups?.join(",") ?? "",
       "x-user-id": backofficeUser.parameters.userId,
-    } as any;
+    };
 
     // call the io-services-cms API and return the response
     const result = await callIoServicesCms(operationId, requestPayload);