+++ apps/io-services-cms-webapp/src/main.ts
@@ -29,6 +29,8 @@ import { getConfigOrThrow } from "./config";
 import { createRequestDeletionHandler } from "./deletor/request-deletion-handler";
 import { createRequestDetailHandler } from "./detailRequestor/request-detail-handler";
 import { createRequestHistoricizationHandler } from "./historicizer/request-historicization-handler";
+import { createRequestServicesHistoryIngestionRetryHandler } from "./ingestion/request-services-history-ingestion-retry-handler";
+import { createRequestServicesLifecycleIngestionRetryHandler } from "./ingestion/request-services-lifecycle-ingestion-retry-handler";
 import { createRequestServicesPublicationIngestionRetryHandler } from "./ingestion/request-services-publication-ingestion-retry-handler";
 import { createServiceTopicIngestorHandler } from "./ingestion/service-topic-ingestor-handler";
 import {
@@ -62,6 +64,8 @@ import { handler as onLegacyServiceChangeHandler } from "./watchers/on-legacy-se
 import { handler as onServiceDetailLifecycleChangeHandler } from "./watchers/on-service-detail-lifecycle-change";
 import { handler as onServiceDetailPublicationChangeHandler } from "./watchers/on-service-detail-publication-change";
 import { handler as onServiceHistoryHandler } from "./watchers/on-service-history-change";
+import { handler as onIngestionServiceHistoryChangeHandler } from "./watchers/on-service-ingestion-history-change";
+import { handler as onIngestionServiceLifecycleChangeHandler } from "./watchers/on-service-ingestion-lifecycle-change";
 import { handler as onIngestionServicePublicationChangeHandler } from "./watchers/on-service-ingestion-publication-change";
 import { handler as onServiceLifecycleChangeHandler } from "./watchers/on-service-lifecycle-change";
 import { handler as onServicePublicationChangeHandler } from "./watchers/on-service-publication-change";
@@ -156,10 +160,23 @@ const servicePublicationEventHubProducer = new EventHubProducerClient(
   config.SERVICES_PUBLICATION_EVENT_HUB_NAME,
 );
 
+
 // eventhub producer for ServiceTopics
 const serviceTopicsEventHubProducer = new EventHubProducerClient(
   config.SERVICES_TOPICS_EVENT_HUB_CONNECTION_STRING,
   config.SERVICES_TOPICS_EVENT_HUB_NAME,
+);
+
+// eventhub producer for ServiceLifecycle
+const serviceLifecycleEventHubProducer = new EventHubProducerClient(
+  config.SERVICES_LIFECYCLE_EVENT_HUB_CONNECTION_STRING,
+  config.SERVICES_LIFECYCLE_EVENT_HUB_NAME,
+);
+
+// eventhub producer for ServiceHistory
+const serviceHistoryEventHubProducer = new EventHubProducerClient(
+  config.SERVICES_HISTORY_EVENT_HUB_CONNECTION_STRING,
+  config.SERVICES_HISTORY_EVENT_HUB_NAME,
 );
 
 // entrypoint for all http functions
@@ -395,3 +412,44 @@ export const serviceTopicIngestorEntryPoint = createServiceTopicIngestorHandler(
   getServiceTopicDao(config),
   serviceTopicsEventHubProducer,
 );
+
+export const onIngestionServiceLifecycleChangeEntryPoint = pipe(
+  onIngestionServiceLifecycleChangeHandler(serviceLifecycleEventHubProducer),
+  processAllOf(ServiceLifecycle.CosmosResource),
+  setBindings((results) => ({
+    ingestionError: pipe(
+      results,
+      RA.map(RR.lookup("ingestionError")),
+      RA.filter(O.isSome),
+      RA.map((item) => pipe(item.value, JSON.stringify)),
+    ),
+  })),
+  toAzureFunctionHandler,
+);
+
+//Ingestion Service History
+export const onIngestionServiceHistoryChangeEntryPoint = pipe(
+  onIngestionServiceHistoryChangeHandler(serviceHistoryEventHubProducer),
+  processAllOf(ServiceHistory),
+  setBindings((results) => ({
+    ingestionError: pipe(
+      results,
+      RA.map(RR.lookup("ingestionError")),
+      RA.filter(O.isSome),
+      RA.map((item) => pipe(item.value, JSON.stringify)),
+    ),
+  })),
+  toAzureFunctionHandler,
+);
+
+//Ingestion Service Lifecycle Retry DLQ
+export const createRequestServicesLifecycleIngestionRetryEntryPoint =
+  createRequestServicesLifecycleIngestionRetryHandler(
+    serviceLifecycleEventHubProducer,
+  );
+//Ingestion Service History Retry DLQ
+export const createRequestServicesHistoryIngestionRetryEntryPoint =
+  createRequestServicesHistoryIngestionRetryHandler(
+    serviceHistoryEventHubProducer,
+  );
+