+++ infra/_modules/cms_function_app/main.tf
@@ -51,6 +51,10 @@ module "cms_fn" {
       "AzureWebJobs.ServiceDetailPublicationWatcher.Disabled"            = "0"
       "AzureWebJobs.IngestionServicePublicationWatcher.Disabled"         = "1"
       "AzureWebJobs.OnRequestServicesPublicationIngestionRetry.Disabled" = "1"
+      "AzureWebJobs.IngestionServiceLifecycleWatcher.Disabled"           = "1"
+      "AzureWebJobs.OnRequestServicesLifecycleIngestionRetry.Disabled"   = "1"
+      "AzureWebJobs.IngestionServiceHistoryWatcher.Disabled"             = "1"
+      "AzureWebJobs.OnRequestServicesHistoryIngestionRetry.Disabled"     = "1"
     }
   )
 
@@ -76,6 +80,10 @@ module "cms_fn" {
       "AzureWebJobs.ServiceDetailPublicationWatcher.Disabled"            = "1"
       "AzureWebJobs.IngestionServicePublicationWatcher.Disabled"         = "1"
       "AzureWebJobs.OnRequestServicesPublicationIngestionRetry.Disabled" = "1"
+      "AzureWebJobs.IngestionServiceLifecycleWatcher.Disabled"           = "1"
+      "AzureWebJobs.OnRequestServicesLifecycleIngestionRetry.Disabled"   = "1"
+      "AzureWebJobs.IngestionServiceHistoryWatcher.Disabled"             = "1"
+      "AzureWebJobs.OnRequestServicesHistoryIngestionRetry.Disabled"     = "1"
     }
   )
 
@@ -98,7 +106,11 @@ module "cms_fn" {
     "AzureWebJobs.ServiceDetailLifecycleWatcher.Disabled",
     "AzureWebJobs.ServiceDetailPublicationWatcher.Disabled",
     "AzureWebJobs.IngestionServicePublicationWatcher.Disabled",
-    "AzureWebJobs.OnRequestServicesPublicationIngestionRetry.Disabled"
+    "AzureWebJobs.OnRequestServicesPublicationIngestionRetry.Disabled",
+    "AzureWebJobs.IngestionServiceLifecycleWatcher.Disabled",
+    "AzureWebJobs.OnRequestServicesLifecycleIngestionRetry.Disabled",
+    "AzureWebJobs.IngestionServiceHistoryWatcher.Disabled",
+    "AzureWebJobs.OnRequestServicesHistoryIngestionRetry.Disabled"
   ]
 
   tier = local.cms.tier